+++ docs/changelog.md
@@ -14,6 +14,7 @@ category: Administration
 * feat: extend syntax and allow to set relative dates (#73)
 * feat: validate for absolute or relative date on due date field (#73)
 * docs: add documentation for relative dates (#73)
+* feat: improve logging and output for template import
 
 ### [23.03.1] - 2023-03-13